adding dom-backed spacebar cycling
