+++ editor/src/components/editor/global-shortcuts.ts
@@ -9,7 +9,7 @@ import {
 } from '../../core/shared/project-file-types'
 import * as PP from '../../core/shared/property-path'
 import * as TP from '../../core/shared/template-path'
-import { CanvasMousePositionRaw } from '../../templates/editor-canvas'
+import { CanvasMousePositionRaw, WindowMousePositionRaw } from '../../templates/editor-canvas'
 import Keyboard, {
   KeyCharacter,
   KeysPressed,
@@ -22,6 +22,7 @@ import Utils from '../../utils/utils'
 import Canvas, { TargetSearchType } from '../canvas/canvas'
 import CanvasActions from '../canvas/canvas-actions'
 import { adjustAllSelectedFrames } from '../canvas/controls/select-mode/move-utils'
+import { getAllTargetsAtPoint } from '../canvas/dom-lookup'
 import {
   toggleBackgroundLayers,
   toggleBorder,
@@ -462,13 +463,7 @@ export function handleKeyDown(
           if (CanvasMousePositionRaw == null) {
             return [EditorActions.clearSelection()]
           }
-          const targetStack = Canvas.getAllTargetsAtPoint(
-            editor,
-            CanvasMousePositionRaw,
-            [TargetSearchType.All],
-            true,
-            'strict', // _IF_ we want to enable loose targeting for selection, it means we also need to change component-area-control
-          )
+          const targetStack = getAllTargetsAtPoint(WindowMousePositionRaw)
           const nextTarget = Canvas.getNextTarget(editor.selectedViews, targetStack)
           if (targetStack.length === 0 || nextTarget === null) {
             return [EditorActions.clearSelection()]
